added interfaces matching implementations
