+++ src/main/java/socialmediaapp/twitterinspiredapp/controller/PostsController.java
@@ -3,7 +3,7 @@ package socialmediaapp.twitterinspiredapp.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import socialmediaapp.twitterinspiredapp.dto.PostDto;
-import socialmediaapp.twitterinspiredapp.service.PostService;
+import socialmediaapp.twitterinspiredapp.service.PostServiceImpl;
 
 import java.util.List;
 
@@ -13,31 +13,31 @@ import static org.springframework.http.HttpStatus.OK;
 @RestController
 @RequestMapping("/api/posts")
 public class PostsController {
-    private final PostService postService;
+    private final PostServiceImpl postServiceImpl;
 
-    public PostsController(PostService postService) {
-        this.postService = postService;
+    public PostsController(PostServiceImpl postServiceImpl) {
+        this.postServiceImpl = postServiceImpl;
     }
 
     @PostMapping
     public ResponseEntity<PostDto> createPost(@RequestBody PostDto postDto) {
-        return new ResponseEntity<>(postService.save(postDto),CREATED);
+        return new ResponseEntity<>(postServiceImpl.save(postDto),CREATED);
     }
 
     @GetMapping("/all")
     public ResponseEntity<List<PostDto>> getAllPosts() {
 
-        return new ResponseEntity<>(postService.getAllPosts(), OK);
+        return new ResponseEntity<>(postServiceImpl.getAllPosts(), OK);
     }
 
     @GetMapping("/all-by-user/{id}")
     public ResponseEntity<List<PostDto>> getAllPostsForUser(@PathVariable long id) {
-        return new ResponseEntity<>(postService.getAllPostsForUser(id), OK);
+        return new ResponseEntity<>(postServiceImpl.getAllPostsForUser(id), OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<PostDto> getPost(@PathVariable Long id){
-        return new ResponseEntity<>(postService.getPostById(id),OK);
+        return new ResponseEntity<>(postServiceImpl.getPost(id),OK);
     }
