+++ src/test/java/socialmediaapp/twitterinspiredapp/controller/PostsControllerTest.java
@@ -4,21 +4,17 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import socialmediaapp.twitterinspiredapp.dto.PostDto;
-import socialmediaapp.twitterinspiredapp.model.Post;
 import socialmediaapp.twitterinspiredapp.model.User;
-import socialmediaapp.twitterinspiredapp.service.PostService;
+import socialmediaapp.twitterinspiredapp.service.PostServiceImpl;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.*;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -32,7 +28,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 public class PostsControllerTest {
 
     @MockBean
-    PostService postService;
+    PostServiceImpl postServiceImpl;
 
     @Autowired
     MockMvc mockMvc;
@@ -45,7 +41,7 @@ public class PostsControllerTest {
     @Test
     public void getAllPosts() throws Exception {
 
-        when(postService.getAllPosts()).thenReturn(Arrays.asList(getPostDto(1), getPostDto(2)));
+        when(postServiceImpl.getAllPosts()).thenReturn(Arrays.asList(getPostDto(1), getPostDto(2)));
 
         mockMvc.perform(get("/api/posts/all"))
                 .andDo(print())
@@ -63,7 +59,7 @@ public class PostsControllerTest {
     public void getPost() throws Exception {
         PostDto dto = getPostDto(1);
 
-        when(postService.getPostById(1L)).thenReturn(dto);
+        when(postServiceImpl.getPost(1L)).thenReturn(dto);
 
         mockMvc.perform(get("/api/posts/1"))
                 .andDo(print())
